virtio-serial: Disable flow control for RHEL 5.0 machine type
RH-Author: Amit Shah <amit.shah@redhat.com>
Message-id: <11770c5023308939b951abca3b3471d73999b3dd.1297062544.git.amit.shah@redhat.com>
Patchwork-id: 17783
O-Subject: [PATCH] virtio-serial: Disable flow control for RHEL 5.0 machine type
Bugzilla: 588916
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Michael S. Tsirkin <mst@redhat.com>

Alex pointed out we'll need to disable flow control for RHEL 5 in
addition to RHEL6.0 machine type.

We need to do this to ensure a RHEL6.1+ qemu started with -M rhel5 can
migrate fine to qemu started on a RHEL5.

Bugzilla: 588916
(riding on top of the flow control bug)

Signed-off-by: Amit Shah <amit.shah@redhat.com>
---
 hw/pc.c |    4 ++++
 1 files changed, 4 insertions(+), 0 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>