vhost: disable on tap link down
RH-Author: Michael S. Tsirkin <mst@redhat.com>
Message-id: <747c489b9712a21e8862b3eff7c147d626d94998.1297262374.git.mst@redhat.com>
Patchwork-id: 17908
O-Subject: [PATCHv2 RHEL6.1 2/2] vhost: disable on tap link down
Bugzilla: 676015
RH-Acked-by: Xiao Wang <jasowang@redhat.com>
RH-Acked-by: Glauber Costa <glommer@redhat.com>
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>

qemu makes it possible to disable link at tap
which is not communicated to the guest but
causes all packets to be dropped.

When vhost-net is enabled, vhost needs to be aware of both the virtio
link_down and the peer link_down. we switch to userspace emulation when
either is down.

Signed-off-by: Michael S. Tsirkin <mst@redhat.com>
Reported-by: pradeep <psuriset@linux.vnet.ibm.com>
---
 hw/virtio-net.c |    3 ++-
 1 files changed, 2 insertions(+), 1 deletions(-)

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>