kvm: x86: Introduce kvmclock device to save/restore its state
RH-Author: Glauber Costa <glommer@redhat.com>
Message-id: <1297274207-9911-1-git-send-email-glommer@redhat.com>
Patchwork-id: 17930
O-Subject: [PATCH RHEL6.1] kvm: x86: Introduce kvmclock device to save/restore its state
Bugzilla: 658467
RH-Acked-by: Michael S. Tsirkin <mst@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>

RH-Author: Glauber Costa
RH-Bugzilla: 658467
RH-Upstream-Status: Posted

If kvmclock is used, which implies the kernel supports it, register a
kvmclock device with the sysbus. Its main purpose is to save and restore
the kernel state on migration, but this will also allow to visualize it
one day.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>
Signed-off-by: Glauber Costa <glommer@redhat.com>
---
Upstream diff: I decided not to register a new qdev device for kvmclock,
in order not to complicate our migration paths. Alternatively, I just got
the code motion to hw/kvmclock.c, and the pre_save function.

We also do not have the CONFIG_KVM_PARA switch in our tree.
---
 Makefile.target |    3 +-
 hw/kvmclock.c   |  102 +++++++++++++++++++++++++++++++++++++++++++++++++++++++
 hw/pc.c         |    3 ++
 qemu-kvm-x86.c  |   35 -------------------
 4 files changed, 107 insertions(+), 36 deletions(-)
 create mode 100644 hw/kvmclock.c

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>