add a service to reap zombies, use it in SLIRP
RH-Author: Paolo Bonzini <pbonzini@redhat.com>
Message-id: <1299775781-3350-1-git-send-email-pbonzini@redhat.com>
Patchwork-id: 19903
O-Subject: [PATCH] add a service to reap zombies, use it in SLIRP
Bugzilla: 678524
RH-Acked-by: Glauber Costa <glommer@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>
RH-Acked-by: Eric Blake <eblake@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=678524

Upstream status: submitted

Brew build: https://brewweb.devel.redhat.com/taskinfo?taskID=3168226

SLIRP -smb support wants to fork a process and forget about reaping it.
To please it, add a generic service to register a process id and let
QEMU reap it.  In the future it could be enhanced to pass a status,
but this would be unused.

With this in place, the SIGCHLD signal handler would not stomp on pclose
anymore.

The issue does not affect RHEL5.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
---
        Posting before approval upstream to get impressions for
        this bug which is a 6.1.0 blocker.  Replies there are
        helpful as well.

	Please review and ack for both 6.1.0 and 6.0.z.

 Makefile      |    2 +-
 iohandler.c   |   92 +++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 qemu-common.h |    1 +
 slirp/misc.c  |    5 ++-
 vl.c          |    9 -----
 5 files changed, 98 insertions(+), 11 deletions(-)
 create mode 100644 iohandler.c

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>