spice-qemu-char.c: remove intermediate buffer
RH-Author: Alon Levy <alevy@redhat.com>
Message-id: <1300789680-18535-3-git-send-email-alevy@redhat.com>
Patchwork-id: 20378
O-Subject: [PATCH RHEL6.1 v2 2/2] spice-qemu-char.c: remove intermediate buffer
Bugzilla: 672191
RH-Acked-by: Amit Shah <amit.shah@redhat.com>
RH-Acked-by: Uri Lublin <uril@redhat.com>
RH-Acked-by: Hans de Goede <hdegoede@redhat.com>

BZ: 672191
upstream: not submitted (explained below)

virtio-serial's buffer is valid when it calls us, and we don't
access it otherwise: vmc_read is only called in response to wakeup,
or else we set datalen=0 and throttle. Then vmc_read is called back,
we return 0 (not accessing the buffer) and set the timer to unthrottle.

Also make datalen int and not ssize_t (to fit spice_chr_write signature).

This relied on the previous patch that introduces throttling, which
can't go upstream right now as explained in that patch.
---
 spice-qemu-char.c |   18 ++++++------------
 1 files changed, 6 insertions(+), 12 deletions(-)

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>