chardev: Allow frontends to notify backends of guest open / close
RH-Author: Hans de Goede <hdegoede@redhat.com>
Message-id: <1300801593-1297-2-git-send-email-hdegoede@redhat.com>
Patchwork-id: 20398
O-Subject: [PATCH 1/3] chardev: Allow frontends to notify backends of guest open
	/ close
Bugzilla: 688572
RH-Acked-by: Uri Lublin <uril@redhat.com>
RH-Acked-by: Amit Shah <amit.shah@redhat.com>
RH-Acked-by: Alon Levy <alevy@redhat.com>

Some frontends know when the guest has opened the "channel" and is actively
listening to it, for example virtio-serial. This patch adds 2 new qemu-chardev
functions which can be used by frontends to signal guest open / close, and
allows interested backends to listen to this.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
---
 qemu-char.c |   18 ++++++++++++++++++
 qemu-char.h |    4 ++++
 2 files changed, 22 insertions(+), 0 deletions(-)

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>