Do not delete BlockDriverState when deleting the drive
RH-Author: Markus Armbruster <armbru@redhat.com>
Message-id: <1301392258-4532-3-git-send-email-armbru@redhat.com>
Patchwork-id: 20729
O-Subject: [PATCH RHEL6.1 qemu-kvm 2/2] Do not delete BlockDriverState when
	deleting the drive
Bugzilla:
RH-Acked-by: Kevin Wolf <kwolf@redhat.com>
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
Bugzilla-related: 654682

This is a backport of Ryan Harper's v4 as seen in my crystal ball (he
hasn't posted it, yet, but the errors I pointed out in v3 are pretty
straightforward to fix).

v3's commit message:

When removing a drive from the host-side via drive_del we currently have the
following path:

drive_del
qemu_aio_flush()
bdrv_close()    // zaps bs->drv, which makes any subsequent I/O get
                // dropped.  Works as designed
drive_uninit()
bdrv_delete()   // frees the bs.  Since the device is still connected to
                // bs, any subsequent I/O is a use-after-free.

The value of bs->drv becomes unpredictable on free.  As long as it
remains null, I/O still gets dropped, however it could become non-null at any
point after the free resulting SEGVs or other QEMU state corruption.

To resolve this issue as simply as possible, we can chose to not actually
delete the BlockDriverState pointer.  Since bdrv_close() handles setting the drv
pointer to NULL, we just need to remove the BlockDriverState from the QLIST
that is used to enumerate the block devices.  This is currently handled within
bdrv_delete, so move this into its own function, bdrv_make_anon().

The result is that we can now invoke drive_del, this closes the file descriptors
and sets BlockDriverState->drv to NULL which prevents futher IO to the device,
and since we do not free BlockDriverState, we don't have to worry about the copy
retained in the block devices.

We also don't attempt to remove the qdev property since we are no longer deleting
the BlockDriverState on drives with associated drives.  This also allows for
removing Drives with no devices associated either.

Reported-by: Markus Armbruster <armbru@redhat.com>
Signed-off-by: Ryan Harper <ryanh@us.ibm.com>
Signed-off-by: Markus Armbruster <armbru@redhat.com>
---
 block.c             |   14 +++++++++++---
 block.h             |    1 +
 hw/device-hotplug.c |   25 ++++++++-----------------
 3 files changed, 20 insertions(+), 20 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>