ide: Split error status from status register
RH-Author: Kevin Wolf <kwolf@redhat.com>
Message-id: <1308664766-10499-2-git-send-email-kwolf@redhat.com>
Patchwork-id: 27715
O-Subject: [RHEL-6.2 qemu-kvm PATCH 1/4] ide: Split error status from status register
Bugzilla: 698537
RH-Acked-by: Markus Armbruster <armbru@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>

Bugzilla: 698537

When adding the werror=stop mode, some flags were added to s->status
which are used to determine what kind of operation should be restarted
when the VM is continued.

Unfortunately, it turns out that s->status is in fact a device register
and as such is visible to the guest (some of the abused bits are even
writable for the guest).

For migration we keep on using the old VMState field (renamed to
migration_compat_status) if the status register doesn't use any of the
previously abused bits. If it does, we use a subsection with a clean copy of
the status register.

The error status is always sent in a subsection if there is any error. It can't
use the old field because errors happen even without PCI.

Signed-off-by: Kevin Wolf <kwolf@redhat.com>
(cherry picked from commit def93791f22a3536c3508244ec7d270098484c7d)

Conflicts:

	hw/ide/core.c
	hw/ide/pci.c
	hw/ide/pci.h

Signed-off-by: Kevin Wolf <kwolf@redhat.com>
---
 hw/ide/core.c     |   46 +++++++++++++++++++++++++++++++++++++------
 hw/ide/internal.h |   12 +++++++++++
 hw/ide/pci.c      |   55 ++++++++++++++++++++++++++++++++++++++++++++++++++++-
 3 files changed, 105 insertions(+), 8 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>