Add an isa device for SGA
RH-Author: Glauber Costa <glommer@redhat.com>
Message-id: <1308752475-27865-1-git-send-email-glommer@redhat.com>
Patchwork-id: 27783
O-Subject: [PATCH RHEL6.2] Add an isa device for SGA
Bugzilla: 684949
RH-Acked-by: Markus Armbruster <armbru@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Luiz Capitulino <lcapitulino@redhat.com>

RH-Author: Glauber Costa <glommer@redhat.com>
RH-Upstream-status: uq/master, a90d4690074526f54ad0851fce19fa6783f06803
RH-Bugzilla: 684949

RH-Notes: For this to work, we need to add a new package in RHEL6. The
new package request is being tracked at #712993. However, there is no
penalty in having the patch early in our tree, since it is only effective
if we enable it using -device sga commandline. And having it early will
allow libvirt people to start working on support for it.

This patch adds a dummy legacy ISA device whose responsibility is to
deploy sgabios, an option rom for a serial graphics adapter.
The proposal is that this device is always-on when -nographics,
but can otherwise be enable in any setup when -device sga is used.

[v2: suggestions on qdev by Markus ]
[v3: cleanups and documentation, per list suggestions ]

Signed-off-by: Glauber Costa <glommer@redhat.com>
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
---
 Makefile.target |    2 +-
 hw/pc.c         |    9 ++++++++
 hw/sga.c        |   56 +++++++++++++++++++++++++++++++++++++++++++++++++++++++
 3 files changed, 66 insertions(+), 1 deletions(-)
 create mode 100644 hw/sga.c

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>