usb: Add a speedmask to devices
RH-Author: Gerd Hoffmann <kraxel@redhat.com>
Message-id: <1311266648-1179-7-git-send-email-kraxel@redhat.com>
Patchwork-id: 30130
O-Subject: [RHEL-6.2 kvm PATCH 06/30] usb: Add a speedmask to devices
Bugzilla: 723858 723863
RH-Acked-by: Hans de Goede <hdegoede@redhat.com>
RH-Acked-by: Kevin Wolf <kwolf@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>

This is used to indicate at which speed[s] the device can operate,
so that this can be checked to match the ports capabilities when it gets
attached to a bus.

Note that currently all usb1 emulated device claim to be fullspeed, this
seems to not cause any problems, but still seems wrong, because with real
hardware keyboards, mice and tablets usually are lo-speed, so reporting these
as fullspeed devices seems wrong.

Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>
(cherry picked from commit ba3f9bfba9ad8d2bbcb9a9fa9ed0f07347fc1c58)

Conflicts:

	usb-bsd.c
---
 hw/usb-ccid.c |    1 +
 hw/usb-desc.c |   10 +++++++
 hw/usb.h      |    3 ++
 usb-bsd.c     |   74 ++++++++++++++++++++++++++++++--------------------------
 usb-linux.c   |    1 +
 5 files changed, 55 insertions(+), 34 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>