usb/bluetooth: compile out the crap
RH-Author: Gerd Hoffmann <kraxel@redhat.com>
Message-id: <1311341959-13308-1-git-send-email-kraxel@redhat.com>
Patchwork-id: 30175
O-Subject: [RHEL-6.2 kvm PATCH v2] usb/bluetooth: compile out the crap
Bugzilla: 723864
RH-Acked-by: Hans de Goede <hdegoede@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

This patch disables a bunch of device emulations which we are not
willing to support.  None of these has libvirt support.

The following usb devices are disabled:
 * usb-wacom-tablet (use usb-tablet instead).
 * usb-braille (useless anyway as we don't build with brlapi).
 * usb-serial
 * usb-net
 * usb-bt-dongle

Together with the bluetooth dongle all other bluetooth emulation code
is disabled too as it is is just dead code without the dongle.

[ v2: change comment to match others in the tree,
      more verbose commit message ]

Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>
---
 Makefile |    6 ++----
 vl.c     |    8 ++++++++
 2 files changed, 10 insertions(+), 4 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>