qed: extract qed_start_allocating_write()
RH-Author: Anthony Liguori <aliguori@redhat.com>
Message-id: <1313163503-2523-6-git-send-email-aliguori@redhat.com>
Patchwork-id: 31328
O-Subject: [RHEL6.2 qemu PATCH 05/15] qed: extract qed_start_allocating_write()
Bugzilla: 633370
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
RH-Acked-by: Kevin Wolf <kwolf@redhat.com>
RH-Acked-by: Orit Wasserman <owasserm@redhat.com>
RH-Acked-by: Christoph Hellwig <chellwig@redhat.com>

From: Stefan Hajnoczi <stefanha@linux.vnet.ibm.com>

Copy-on-read requests are a form of allocating write and will need to be
queued like other allocating writes.  This patch extracts the request
queuing code for allocating writes so that it can be reused for
copy-on-read in a later patch.

Signed-off-by: Stefan Hajnoczi <stefanha@linux.vnet.ibm.com>
Signed-off-by: Anthony Liguori <aliguori@redhat.com>

Bugzilla: 633370

---
 block/qed.c |   32 ++++++++++++++++++++++++++------
 1 files changed, 26 insertions(+), 6 deletions(-)

Signed-off-by: Michal Novotny <mignov@gmail.com>