qed: add support for copy-on-read
RH-Author: Anthony Liguori <aliguori@redhat.com>
Message-id: <1313163503-2523-8-git-send-email-aliguori@redhat.com>
Patchwork-id: 31333
O-Subject: [RHEL6.2 qemu PATCH 07/15] qed: add support for copy-on-read
Bugzilla: 633370
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
RH-Acked-by: Kevin Wolf <kwolf@redhat.com>
RH-Acked-by: Orit Wasserman <owasserm@redhat.com>
RH-Acked-by: Christoph Hellwig <chellwig@redhat.com>

From: Anthony Liguori <aliguori@us.ibm.com>

This patch implements copy-on-read in QED.  Once a read request reaches
the copy-on-read state it adds itself to the allocating write queue in
order to avoid race conditions with write requests.

If an allocating write request manages to sneak in before the
copy-on-read request, then the copy-on-read will notice that the cluster
has been allocated when qed_find_cluster() is re-run.  This works
because only one allocating request is active at any time and when the
next request is activated it will re-run qed_find_cluster().

[Originally by Anthony.  Stefan added allocating write queuing and
factored out the QED_CF_COPY_ON_READ header flag.]

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
Signed-off-by: Stefan Hajnoczi <stefanha@linux.vnet.ibm.com>
Signed-off-by: Anthony Liguori <aliguori@redhat.com>

Bugzilla: 633370

---
 block/qed.c  |   35 +++++++++++++++++++++++++++++++++--
 block/qed.h  |    3 ++-
 trace-events |    1 +
 3 files changed, 36 insertions(+), 3 deletions(-)

Signed-off-by: Michal Novotny <mignov@gmail.com>