bz716261: Enable +xsave as a -cpu flag
RH-Author: ddugger@redhat.com
Message-id: <20110922172349.GB17560@sobek.n0ano.com>
Patchwork-id: 33204
O-Subject: Re: [RHEL 6.2 PATCH 7/6 V3] bz716261: Enable +xsave as a -cpu flag
Bugzilla: 716261
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
RH-Acked-by: Eduardo Habkost <ehabkost@redhat.com>
RH-Acked-by: Avi Kivity <avi@redhat.com>

Allows the user to specify the cpu type `-cpu +xsave' to expose just
the xsave CPUID capability to the guest.

Although not specifically required for xsave support this is an extremely
nice to have feature that requires a very simple change and would
therefore be appropriate for adding to this patch series.

Testing: Has been tested and does properly enable the XSAVE capability.

Signed-off-by: Don Dugger <donald.d.dugger@intel.com>
---
 target-i386/helper.c |    2 +-
 1 files changed, 1 insertions(+), 1 deletions(-)

Signed-off-by: Michal Novotny <minovotn@redhat.com>