Fix packaging of the QEMU guest agent
The binary 'qemu-ga' is not a userspace emulator, it is the
QEMU guest agent. Since this typically needs to be installed
in the guest put it in a sub-RPM.

A systemd service is used to start the agent. A udev rule is
used to automatically start the systemd service, if the
corresponding virtio-serial port is present