runstate: Allow user to migrate twice
RH-Author: Luiz Capitulino <lcapitulino@redhat.com>
Message-id: <1319036039-4358-13-git-send-email-lcapitulino@redhat.com>
Patchwork-id: 34416
O-Subject: [PATCH RHEL6.2 qemu-kvm v3 12/13] runstate: Allow user to migrate twice
Bugzilla: 617889
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>
RH-Acked-by: Kevin Wolf <kwolf@redhat.com>

It should be a matter of allowing the transition POSTMIGRATE ->
FINISH_MIGRATE, but it turns out that the VM won't do the
transition the second time because it's already stopped.

So this commit also adds vm_stop_force_state() which performs
the transition even if the VM is already stopped.

While there also allow other states to migrate.

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>

Conflicts:

	cpus.c
	migration.c
	vl.c

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>
---
 migration.c |    2 +-
 sysemu.h    |    1 +
 vl.c        |   21 +++++++++++++++++++--
 3 files changed, 21 insertions(+), 3 deletions(-)

Signed-off-by: Michal Novotny <minovotn@redhat.com>