RunState: Don't abort on invalid transitions
RH-Author: Luiz Capitulino <lcapitulino@redhat.com>
Message-id: <1319036039-4358-14-git-send-email-lcapitulino@redhat.com>
Patchwork-id: 34415
O-Subject: [PATCH RHEL6.2 qemu-kvm v3 13/13] RunState: Don't abort on invalid transitions
Bugzilla: 617889
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>
RH-Acked-by: Kevin Wolf <kwolf@redhat.com>

This commit disables the abort() call which is triggered on invalid
transitions. After discussing this with Markus we concluded that
we're not confident enough that the current transition table captures
all the possible valid transions.

This commit also changes runstate_set() to print a nice warning
message when an invalid transition is made.

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>
---
 vl.c |   13 +++++++++----
 1 files changed, 9 insertions(+), 4 deletions(-)

Signed-off-by: Michal Novotny <minovotn@redhat.com>