char: Disable write callback if throttled chardev is detached
RH-Author: Amit Shah <amit.shah@redhat.com>
Message-id: <504456631697d43d20eb5207046dcfa1338f0c09.1322820758.git.amit.shah@redhat.com>
Patchwork-id: 35531
O-Subject: [RHEL6.3 qemu PATCH 1/1] char: Disable write callback if throttled chardev is detached
Bugzilla: 745758
RH-Acked-by: Hans de Goede <hdegoede@redhat.com>
RH-Acked-by: Jeffrey Cody <jcody@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>

If a frontend device releases the chardev, the chr handlers were set
to NULL via qemu_add_handlers(chr, NULL, NULL).  However, if the
chardev was throttled from the host side, the write_unblocked()
callback will be invoked once the chardev becomes writable, which
continues to use the now NULL handlers.  Ensure the callbacks are
disabled when frontends release chardevs.

This was seen when a virtio-serial port was unplugged when heavy
guest->host IO was in progress.  In the window in which the throttling
was active, unplugging ports caused a qemu segfault.

Bugzilla: 745758
Upstream: Not applicable, since throttling is a RHEL6-only feature.

Signed-off-by: Amit Shah <amit.shah@redhat.com>
Signed-off-by: Michal Novotny <minovotn@redhat.com>